MAJOR CLEANUP.. Unit tests need some work though!
All N, K, M, S parameters are declared in sampler.c as `const`.
These parameters are basically constants of a simulation. Therefore I have
decided to make them proper constants. This has the advantage of having less
cluttered function calls, i.e. no need to pass these parameters around.
Although they are `global`, they can't be mutated. So no need to get
freaked out...
Also compilers like constants for optimization purposes.

The drawback is that, in order to change these parameters, we need to recompile
the code. But this is not that big of a deal in practice, because compilation takes
only a second.
Unit tests need some work..